requirements.txt
black==23.3.0
click==8.1.3
colorama==0.4.6
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
protobuf==3.20.3
python-dotenv==1.0.0
tomli==2.0.1
